services/enricher/requirements.txt
openai
sentence-transformers
pydantic
python-dotenv
azure-search-documents
